requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
astrapy==0.7.7
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
blis==0.7.11
cassandra-driver==3.29.0
cassio==0.1.5
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.1
coverage==7.4.1
cryptography==41.0.7
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.3
deep-translator==1.11.4
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
Faker==19.11.1
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
futures==2.2.0
gensim==4.3.2
geomet==0.2.1.post1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hstspreload==2024.2.1
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.22.2
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.12
langchain-astradb==0.0.1
langchain-community==0.0.28
langchain-core==0.1.50
langchain-experimental==0.0.54
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langchainhub==0.1.14
langcodes==3.3.0
langsmith==0.1.25
lingua-language-detector==2.0.2
lxml==4.9.4
Markdown==3.5.1
markdownify==0.11.6
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.4
memory-profiler==0.61.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==1.24.2
openai==1.25.2
orjson==3.10.1
packaging==23.2
pandas==2.1.4
pdf2image==1.17.0
pdfminer.six==20231228
phonenumbers==8.13.30
platformdirs==4.2.1
pluggy==1.4.0
preshed==3.0.9
presidio-analyzer==2.2.353
presidio-anonymizer==2.2.353
psutil==5.9.8
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
Pympler==1.0.1
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.1.2
pypdf==3.17.4
pytesseract==0.3.10
pytest==7.4.4
pytest-cov==4.1.0
pytest-testdox==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jobspy==1.1.52
pytz==2023.3.post1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
srsly==2.4.8
sympy==1.12
tenacity==8.2.3
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
tldextract==5.1.1
tls-client==1.0.1
toml==0.10.2
tomli==2.0.1
typer==0.9.0
types-requests==2.31.0.20231231
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
